+++ tests/preferences/widgets/__init__.py
@@ -1,0 +1,1 @@
+"""Unittests for the preferences widgets."""
